extract packagist tests into separate module file
